improve: update database

+++ alembic/versions/8d1de7cb03de_add_new_columns_to_admins_table.py
@@ -5,35 +5,50 @@ Revises:
 Create Date: 2025-02-24 12:24:43.250096
 
 """
-from typing import Sequence, Union
-
 from alembic import op
 import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '8d1de7cb03de'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
+from sqlalchemy import inspect
 
 
 def upgrade() -> None:
+    inspector = inspect(op.get_bind())
 
-    op.add_column('admins', sa.Column('debt', sa.Integer(), server_default=sa.text("0")))
-    op.add_column("settings", sa.Column("debt_system", sa.Boolean(), server_default=sa.text("0")))
-    
-    op.add_column('admins', sa.Column('traffic_new', sa.String()))
-    op.execute('UPDATE admins SET traffic_new = traffic')
-    op.drop_column('admins', 'traffic')
-    op.alter_column('admins', 'traffic_new', new_column_name='traffic')
+    columns_admins = inspector.get_columns('admins')
+    column_names_admins = [column['name'] for column in columns_admins]
+
+    if 'debt' not in column_names_admins:
+        op.add_column('admins', sa.Column('debt', sa.Integer(), server_default=sa.text("0")))
+
+    if 'traffic_new' not in column_names_admins:
+        op.add_column('admins', sa.Column('traffic_new', sa.String()))
+        op.execute('UPDATE admins SET traffic_new = traffic')
+        op.drop_column('admins', 'traffic')
+        op.alter_column('admins', 'traffic_new', new_column_name='traffic')
+
+    columns_settings = inspector.get_columns('settings')
+    column_names_settings = [column['name'] for column in columns_settings]
+
+    if 'debt_system' not in column_names_settings:
+        op.add_column("settings", sa.Column("debt_system", sa.Boolean(), server_default=sa.text("0")))
+
 
 def downgrade() -> None:
-    
-    op.drop_column('admins', 'debt')
-    op.drop_column("settings", "debt_system")
+    inspector = inspect(op.get_bind())
 
-    op.add_column('admins', sa.Column('traffic_old', sa.Integer()))
-    op.execute('UPDATE admins SET traffic_old = traffic')
-    op.drop_column('admins', 'traffic')
-    op.alter_column('admins', 'traffic_old', new_column_name='traffic')
+    columns_admins = inspector.get_columns('admins')
+    column_names_admins = [column['name'] for column in columns_admins]
+
+    if 'debt' in column_names_admins:
+        op.drop_column('admins', 'debt')
+
+    columns_settings = inspector.get_columns('settings')
+    column_names_settings = [column['name'] for column in columns_settings]
+
+    if 'debt_system' in column_names_settings:
+        op.drop_column("settings", "debt_system")
+
+    if 'traffic' in column_names_admins:
+        op.add_column('admins', sa.Column('traffic_old', sa.Integer()))
+        op.execute('UPDATE admins SET traffic_old = traffic')
+        op.drop_column('admins', 'traffic')
+        op.alter_column('admins', 'traffic_old', new_column_name='traffic')
